backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-allauth
    #   django-cors-headers
attrs==25.1.0
    # via
    #   jsonschema
    #   referencing
blessed==1.20.0
    # via -r requirements.in
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
crispy-tailwind==1.0.3
    # via -r requirements.in
cryptography==44.0.2
    # via pyjwt
diff-match-patch==20241021
    # via django-import-export
dj-database-url==2.3.0
    # via -r requirements.in
dj-rest-auth==7.0.1
    # via -r requirements.in
django==4.2.20
    # via
    #   -r requirements.in
    #   crispy-tailwind
    #   dj-database-url
    #   dj-rest-auth
    #   django-allauth
    #   django-cors-headers
    #   django-crispy-forms
    #   django-import-export
    #   django-jazzmin
    #   django-multiselectfield
    #   django-picklefield
    #   django-q2
    #   django-recaptcha
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-allauth==65.4.1
    # via dj-rest-auth
django-cors-headers==4.7.0
    # via -r requirements.in
django-crispy-forms==2.3
    # via
    #   -r requirements.in
    #   crispy-tailwind
django-environ==0.12.0
    # via -r requirements.in
django-impersonate==1.9.4
    # via -r requirements.in
django-import-export==4.3.7
    # via -r requirements.in
django-jazzmin==3.0.1
    # via -r requirements.in
django-multiselectfield==0.1.13
    # via -r requirements.in
django-picklefield==3.2
    # via django-q2
django-q2==1.7.6
    # via -r requirements.in
django-recaptcha==4.0.0
    # via -r requirements.in
django-storages==1.14.5
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.0
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
gevent==24.11.1
    # via -r requirements.in
greenlet==3.1.1
    # via gevent
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
numpy==2.2.3
    # via pandas
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.2
    # via gunicorn
pandas==2.2.3
    # via -r requirements.in
psycopg2==2.9.10
    # via -r requirements.in
pycparser==2.22
    # via cffi
pyjwt==2.9.0
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via django-allauth
rpds-py==0.23.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via
    #   blessed
    #   python-dateutil
sqlparse==0.5.3
    # via django
supervisor==4.2.5
    # via -r requirements.in
tablib==3.8.0
    # via django-import-export
typing-extensions==4.12.2
    # via
    #   dj-database-url
    #   referencing
tzdata==2025.1
    # via pandas
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.3.0
    # via requests
wcwidth==0.2.13
    # via blessed
whitenoise==6.9.0
    # via -r requirements.in
xlsxwriter==3.2.2
    # via -r requirements.in
zope-event==5.0
    # via gevent
zope-interface==7.2
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
